🏷️rename: LogRecordAlertModal
