+++ src/components/modalViews/index.ts
@@ -1,2 +1,3 @@
 export { default as TestModal } from "./TestModal";
-export { default as ConfirmModal } from "./ConfirmModal/ConfirmModal";
+export { default as LogRecordConfirmModal } from "./LogRecordConfirmModal/LogRecordConfirmModal";
+export { default as LogRecordAlertModal } from "./LogRecordAlertModal/LogRecordAlertModal";
